Bug 515665: [Target Explorer] Refresh after CANCELING download
After CANCELING a large download, the target folder is now still
refreshed. Also, in case the target folder is just a "linked"
resource or there are multiple aliases holding the target folder,
all those aliases are refreshed.

One loophole is NOT fixed with this: Downloading to folder
/base , havign the download fill /base/sub and refreshing
sub when that's an Eclipse project or linked resource.
In this case, "sub" is *not* refreshed but we can live with
this special case.

Change-Id: I1ea146eed33c28fe67d75b5205a8b4deb6690264
Signed-off-by: ashi <Aijun.Shi@windriver.com>