TCF Debugger: simple initial implementation of TCF reset command
Adds a dropdown button for to reset the currently selection context
in the debug view. The default action is to use the first available
reset. The available resets are displayed in the dropdown list next
to the command in the toolbar.

Right now the reset functionality is only available if at most a
single context is selected in the debug view.

Change-Id: I50c5f76745b990aead63570bd983a940bdc39a8a
Signed-off-by: Sanimir Agovic <sanimir@subpath.org>